perf: int32 local coordinates to cut array memory ~32%
Local sequence coordinates (start/end, and the four g1/g2 position fields in
PAIRWISE_DTYPE) were int64 but are bounded by sequence length, which fits int32
(max ~2.1 Gb). Narrow them to int32 across GENOME_POS_DTYPE, HITS_DTYPE, and
PAIRWISE_DTYPE. Genome-global offsets stay int64, so multi-Gb genomes are
unaffected.

Per-row: GENOME_POS 31->23 (-26%), HITS 33->25 (-24%), PAIRWISE 42->26 (-38%).
At 15 genomes x ~1M SCMs + 15M hits + 30 cached pairs that's ~2.22 -> 1.50 GB
of array memory (0.72 GB / 32% saved). derive_pair also got slightly faster
(less memory bandwidth); detect_blocks unchanged.

Guard: GenomeStore.load now rejects any sequence longer than MAX_SEQUENCE_LENGTH
(2^31-1) with a clear message, so the narrowing can never silently truncate.

214 passed (+2 guard tests); ruff + mypy --strict clean.

+++ syntrack/derive/pair.py
@@ -10,15 +10,17 @@ import numpy as np
 if TYPE_CHECKING:
     from syntrack.store.scm import SCMStore
 
+# Positions are *local* sequence coordinates (int32, guarded at load by
+# MAX_SEQUENCE_LENGTH ~2.1 Gb); no genome-global offset is carried here.
 PAIRWISE_DTYPE = np.dtype(
     [
         ("scm_id_idx", np.int32),
         ("g1_seq_idx", np.int16),
         ("g2_seq_idx", np.int16),
-        ("g1_start", np.int64),
-        ("g1_end", np.int64),
-        ("g2_start", np.int64),
-        ("g2_end", np.int64),
+        ("g1_start", np.int32),
+        ("g1_end", np.int32),
+        ("g2_start", np.int32),
+        ("g2_end", np.int32),
         ("g1_strand", np.int8),
         ("g2_strand", np.int8),
     ]
